Refactor into a single CLI module
* Refactor deposit.py and listen.py into a single CLI module
* Update corresponding tests
* Update entrypoint in Pipfile based on refactor
