+++ Pipfile
@@ -24,5 +24,4 @@ moto = "*"
 python_version = "3.9"
 
 [scripts]
-deposit = "python -c \"from awd.deposit import deposit; deposit()\""
-listen = "python -c \"from awd.listen import listen; listen()\""
+awd  = "python -c \"from awd.cli import cli; cli()\""
